SCEE/requirements.txt
# Comunicación y Asincronismo
# (asyncio viene en la librería estándar de Python)

# Tareas Distribuidas
celery
redis

# Base de Datos y Variables de Entorno
#mysql-connector-python # Tengo dudas sobre usar el flask_alchemy o el mysql-connector-python, pero creo que el segundo es más directo para lo que necesito
python-dotenv
aiomysql

# Extras para desarrollo
black  # Formateador de código